refactor(userspace-dp): extract afxdp/ha session import ownership
Move the synced-session import/delete ownership group out of
afxdp/ha/mod.rs into a dedicated `session_import` submodule (#6247). Pure
code motion; the ordered-apply import invariant is preserved byte-for-byte.

Relocated inherent methods:
- synced_import_cap (aggregate 2N entry ceiling)
- upsert_synced_session (older-generation reject; aggregate admission +
  optional synthesized reverse; forward canonical/alias/index publish;
  forward DNAT + eligible live-session BPF projections; reverse-prewarm
  ownership; independent reverse publish/replace decision; pair-adjacent
  per-worker fanout)
- delete_synced_session / delete_synced_session_gen (generation-guarded
  teardown, DNAT release, reverse companion removal)
- test_install_local_forward_session (#4054 test seam)

No shared-session/index mutex is held across BPF syscalls or worker-queue
acquisition — unchanged. Continue-on-BPF-error counters, generation-zero
compatibility, drop-newest cap, and exact status text are untouched.
